Refactor Documents API endpoints and update .bru scripts
- Move document admin endpoints to Admin namespace and folder
- Introduce DocumentsEndpoints for centralized endpoint mapping
- Update endpoint mapping logic and extension methods
- Add OpenAPI tags and use module name constants
- Rewrite .bru files in structured YAML/JSON format
- Update README and directory structure for new layout
- Improve maintainability and API documentation consistency
